Move shipping services component into shipping directory.
